Renamed file data_classes.py to metadata.py and removed its imports from __init__.py
