+++ tools/pyclient/src/molgenis_emx2_pyclient/__init__.py
@@ -1,2 +1,1 @@
 from .client import Client
-from .data_classes import Column, Table, Schema
