+++ tools/pyclient/src/molgenis_emx2_pyclient/client.py
@@ -9,7 +9,7 @@ import requests
 
 from . import graphql_queries as queries
 from . import utils
-from .data_classes import Schema
+from .metadata import Schema
 from .exceptions import (NoSuchSchemaException, ServiceUnavailableError, SigninError,
                          ServerNotFoundError, PyclientException, NoSuchTableException,
                          NoContextManagerException, GraphQLException, InvalidTokenException,
